# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TAP Admission API
  version: 1.0.0
  description: |
    API REST para el examen de admision TAP. Incluye autenticacion, recuperacion
    de credenciales, administracion de productos, usuarios, perfiles, exportes
    PDF/Excel y control de acceso por secciones asignadas al perfil.
servers:
  - url: http://tap-admission-api.test/api
    description: Local con Laragon
  - url: https://tap-admission-api.onrender.com/api
    description: Produccion en Render
...
paths:
...
  /recover-password:
    post:
      tags: [Auth]
      summary: Recuperar credenciales
      description: Valida que el usuario exista y envia una contrasena temporal al correo registrado.
      operationId: recoverPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverPasswordRequest"
            example:
              email: admin@tap.local
      responses:
        "200":
          description: Correo de recuperacion enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Se enviaron credenciales temporales al correo registrado.
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /refresh-token:
    post:
      tags: [Auth]
      summary: Refrescar token de sesion
      description: Genera un token nuevo para el usuario autenticado y revoca el token anterior.
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token de acceso renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
...
  responses:
    Unauthorized:
      description: Token ausente o invalido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Unauthenticated.
...
    ValidationError:
      description: Error de validacion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
          example:
            message: The given data was invalid.
            errors:
              email:
                - Las credenciales no son validas.
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    RecoverPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: Token Bearer de Laravel Sanctum.
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      required: [id, user_code, name, email, sections]
      properties:
        id:
          type: string
        user_code:
          type: string
          example: USR-ADMIN
        name:
          type: string
          example: Administrador TAP
        email:
          type: string
          format: email
          example: admin@tap.local
        sections:
          type: array
          items:
            type: string
            enum: [products, users, profiles, audit_logs]
...